Put SerializeStream into separate file
